auth_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose==3.3.0
python-multipart==0.0.6
python-jose==3.3.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
